refactor: split parser util functions out into a separate file
