+++ src/linux/launchers/steam.rs
@@ -15,9 +15,8 @@ use tracing::{debug, error, trace, warn};
 
 use crate::{
     data::{Game, GamesParsingError, GamesResult, GamesSlice, Launcher, SupportedLaunchers},
-    utils::{
-        clean_game_title, parse_double_quoted_value, parse_until_key, some_if_dir, some_if_file,
-    },
+    parsers::{parse_double_quoted_value, parse_until_key},
+    utils::{clean_game_title, some_if_dir, some_if_file},
 };
 
 struct ParsableManifestData {
